navbev/perception/backend/requirements.txt
tqdm==4.42.1
torchtext==0.6.0
torchvision==0.8.2
psutil==5.7.2
## referit==1.0.0.dev0
torch==1.7.1
matplotlib==3.1.3
nvidia_ml_py3==7.352.0
scikit_image==0.17.2
nvidia_smi==0.1.3
Pillow==8.1.0
#skimage==0.0
